+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 	- `MySQL`
 	- `Postgres`
 * Now possible to define custom input/output value objects for the query builder.
+* Added a `EventStream` response sender to simplify sending [sever-sent events](https://developer.mozilla.org/en-US/docs/Web/API/Server-sent_events/Using_server-sent_events).
 
 --------------------------------------------------------